seperate redux feature folder to three sub folder.
