+++ src/redux/configureStore.js
@@ -1,8 +1,8 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import logger from 'redux-logger';
 import thunk from 'redux-thunk';
-import rocketsReducer from './rockets/rockets';
-import missionsReducer from './missions/missions';
+import rocketsReducer from './rockets/reducer/reducer';
+import missionsReducer from './missions/reducer/reducer';
 
 const reducer = combineReducers({
   rockets: rocketsReducer,
