+++ src/redux/rockets/actions/actions.js
@@ -1,0 +1,3 @@
+export const GET_ROCKETS = 'space-travel/rockets/GET_ROCKETS';
+export const BOOK_ROCKET = 'space-travel/rockets/BOOK_ROCKET';
+export const CANCEL_BOOK_ROCKET = 'space-travel/rockets/CANCEL_BOOK_ROCKET';
