+++ src/redux/rockets/reducer/reducer.js
@@ -1,0 +1,24 @@
+import { GET_ROCKETS, BOOK_ROCKET, CANCEL_BOOK_ROCKET } from '../actions/actions';
+
+const initialState = [];
+
+const reducer = (state = initialState, action) => {
+  switch (action.type) {
+    case GET_ROCKETS:
+      return action.payload;
+    case BOOK_ROCKET:
+      return state.map((rocket) => {
+        if (rocket.id !== action.payload) return rocket;
+        return { ...rocket, reserved: true };
+      });
+    case CANCEL_BOOK_ROCKET:
+      return state.map((rocket) => {
+        if (rocket.id !== action.payload) return rocket;
+        return { ...rocket, reserved: false };
+      });
+    default:
+      return state;
+  }
+};
+
+export default reducer;
